Changed Flightservice and controller
Made some new changes on controller methods, I also add some new methods with new resource path.

I changed also methods in the flightservice

+++ src/main/java/nl/hu/prbed/airline/airline/application/FlightService.java
@@ -1,22 +1,19 @@
 package nl.hu.prbed.airline.airline.application;
 
+import nl.hu.prbed.airline.airline.application.exception.FlightNotFoundException;
+import nl.hu.prbed.airline.airline.application.exception.InvalidDTOException;
 import nl.hu.prbed.airline.airline.data.FlightRepository;
-import nl.hu.prbed.airline.airline.data.PlaneRepository;
 import nl.hu.prbed.airline.airline.domain.Booking;
 import nl.hu.prbed.airline.airline.domain.Flight;
 import nl.hu.prbed.airline.airline.domain.FlightRoute;
 import nl.hu.prbed.airline.airline.domain.Plane;
-import nl.hu.prbed.airline.airline.presentation.dto.BookingDTO;
-import nl.hu.prbed.airline.airline.presentation.dto.FlightDTO;
-import nl.hu.prbed.airline.airline.presentation.dto.FlightRouteDTO;
-import nl.hu.prbed.airline.airline.presentation.dto.PlaneDTO;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import nl.hu.prbed.airline.airline.presentation.dto.Flight.FlightBookingDTO;
+import nl.hu.prbed.airline.airline.presentation.dto.Flight.FlightDTO;
+import nl.hu.prbed.airline.airline.presentation.dto.Flight.FlightDepartureRouteDTO;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 import javax.transaction.Transactional;
-import java.time.LocalDateTime;
+import java.sql.Date;
 import java.util.List;
 
 //laten we de afweging maken welke onderstaande methodes verwijderd kunnen worden, sommige staan dubbelzinnig zoals bv verwijderen van een vlucht
@@ -39,6 +36,10 @@ public class FlightService {
     }
 
     public Flight createFlight(FlightDTO flightDTO) {
+        if(flightDTO.flightRouteId == null || flightDTO.departureTime == null || flightDTO.planeId == null){
+            throw new InvalidDTOException("Missing some input variables to send!");
+        }
+
         Plane plane = planeService.getPlane(flightDTO.planeId);
         FlightRoute flightRoute = flightRouteService.findFlightRouteByID(flightDTO.flightRouteId);
         return flightRepository.save(new Flight(flightDTO.departureTime, flightRoute, plane));
@@ -48,34 +49,67 @@ public class FlightService {
         return flightRepository.findAll();
     }
 
-    public List<Flight> findFlightsByDeparture(LocalDateTime departure) {
+    public Flight findFlightById(Long id) {
+        if(id == null){
+            throw new InvalidDTOException("No id specified!");
+        }
+        return flightRepository.findById(id).orElseThrow(FlightNotFoundException::new);
+
+    }
+
+    public List<Flight> findFlightsByDeparture(Date departure) {
+        if(departure == null){
+            throw new InvalidDTOException("No departure specified");
+        }
         return flightRepository.findAllByDepartureTime(departure);
     }
 
-    public Flight findFlightRouteAndDeparture(FlightDTO flightDTO) {
-        FlightRoute flightRoute = flightRouteService.findFlightRouteByID(flightDTO.planeId);
-        return flightRepository.findByRouteAndDepartureTime(flightRoute, flightDTO.departureTime).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+    public Flight findFlightRouteAndDeparture(FlightDepartureRouteDTO flightDepartureRoute) {
+        if(flightDepartureRoute == null){
+            throw new InvalidDTOException("No id specified!");
+        }
+        FlightRoute flightRoute = flightRouteService.findFlightRouteByID(flightDepartureRoute.flightRouteId);
+        return flightRepository.findByRouteAndDepartureTime(flightRoute, flightDepartureRoute.departureTime).orElseThrow(FlightNotFoundException::new);
     }
 
-    public Flight updateFlight(FlightDTO flightDTO, PlaneDTO planeDTO, FlightRouteDTO flightRouteDTO) {
-        planeService.updatePlane(planeDTO);
-        flightRouteService.updateFlightRoute(flightRouteDTO);
+    public Flight updateFlight(FlightDTO flightDTO, Long id) {
+        if(id == null){
+            throw new InvalidDTOException("No id specified!");
+        }
 
-        Flight flight = flightRepository.findById(this.flight.getId()).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Flight flight = findFlightById(id);
+        FlightRoute flightRoute = flightRouteService.findFlightRouteByID(flightDTO.flightRouteId);
+        Plane plane = planeService.getPlane(flightDTO.planeId);
+        flight.update(flightDTO.departureTime, flightRoute, plane);
         flightRepository.save(flight);
         return flight;
     }
 
-    public void deleteFlightByRouteAndDeparture(FlightRoute route, LocalDateTime departure) {
+    public void deleteFlightByRouteAndDeparture(FlightRoute route, Date departure) {
+        if(flightRepository.findByRouteAndDepartureTime(route, departure).isEmpty()){
+            throw new FlightNotFoundException();
+        }
         flightRepository.deleteByDepartureTimeAndRoute(departure, route);
     }
 
     public void deleteFlightById(Long id) {
+        if(id == null){
+            throw new InvalidDTOException("No id specified!");
+        }
+        else if(flightRepository.findById(id).isEmpty()){
+            throw new FlightNotFoundException();
+        }
         flightRepository.deleteById(id);
     }
 
-    public void addBooking(BookingDTO bookingDTO) {
-        flight.addBooking(bookingService.createBooking(bookingDTO));
+    public void addBooking(FlightBookingDTO flightBooking) {
+        if(flightBooking.bookingId == null || flightBooking.flightId == null){
+            throw new InvalidDTOException("Missing input variables to send!");
+        }
+        Booking booking = bookingService.findBookingById(flightBooking.bookingId);
+        Flight flight = findFlightById(flightBooking.flightId);
+        flight.addBooking(booking);
         flightRepository.save(flight);
     }
+
 }
